scripts/requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
cohere>=4.0.0
Pillow>=9.0.0
requests>=2.31.0
aiohttp>=3.9.0
mdx2>=0.1.0
python-dotenv>=1.0.0 